test(08.5-11): Q12b mixed-shape generators + uint16 overflow fix
- Q12b mixed-shape testing/quick generator. New helper
  mixedShapeStringGenerator() draws from a balanced five-shape mix
  (20% each):
    - empty string
    - short ASCII (length 1..8, printable ASCII)
    - long non-ASCII (length 64..128, Latin Extended + Greek +
      Cyrillic + CJK + emoji base)
    - all-same-character runs (length 1..32, single ASCII letter)
    - Unicode-rich (length 4..32, diacritics + CJK + emoji +
      combining marks attached to base ASCII)
  Documented as the docs/requirements.md §15.3 testing/quick generator
  contract; locked 2026-05-17 per CONTEXT.md Q12b.

- New TestProp_LevenshteinScore_RangeBounds_MixedShape exercises the
  generator across the [0,1] range-bounds invariant. This is the
  load-bearing "generator must be wired in" gate; other property tests
  deliberately continue to use rand.String to avoid re-validating
  seed sufficiency — broader adoption is Phase 9+ retrofit.

- Q12b uint16 overflow flake fix in TestProp_Scorer_WeightSumOne.
  Previous toPositive(u uint16) used `float64(u+1)` which wraps to 0
  when u == 65535 (uint16 addition saturates inside the type before
  the float64 conversion), producing weight == 0 — rejected by
  ErrInvalidWeight at the option layer as a single-seed flake.
  Replaced with `float64(uint32(u)+1)`: widen BEFORE the +1 so the
  addition cannot wrap. Verified 10 consecutive race runs green.

+++ scorer_test.go
@@ -824,7 +824,15 @@ func TestProp_Scorer_WeightSumOne(t *testing.T) {
 		toPositive := func(u uint16) float64 {
 			// Avoid zero by adding 1 in the numerator; range becomes
 			// (0, 100].
-			return float64(u+1) / float64(uint32(1)<<16) * 100.0
+			//
+			// Q12b uint16 overflow fix: widen u to uint32 BEFORE the
+			// `+ 1` so the addition cannot wrap. The previous form
+			// `float64(u+1)` wrapped to 0 when u == 65535 (the uint16
+			// addition saturates inside the uint16 type before the
+			// float64 conversion), producing weight == 0 which is
+			// rejected by ErrInvalidWeight at the option layer — a
+			// flaky failure on the single uint16-max seed.
+			return float64(uint32(u)+1) / float64(uint32(1)<<16) * 100.0
 		}
 		w := []float64{toPositive(w0), toPositive(w1), toPositive(w2)}
 		algos := []fuzzymatch.AlgoID{
